cleaning, change folder structure
