--- src/views/studio/InitialStudioView/index.ts
@@ -1,3 +1,0 @@
-import InitialStudioView from './InitialStudioView'
-
-export default InitialStudioView
