+++ src/views/studio/SignInJoinView/SignInMainView/index.ts
@@ -1,0 +1,3 @@
+import SignInMainView from './SignInMainView'
+
+export default SignInMainView
