+++ src/views/studio/SignInJoinView/SignInProccessView/SignInProccessView.style.ts
