+++ src/views/studio/SignInJoinView/Steps/AccountStep.style.ts
