+++ src/views/studio/SignInJoinView/Steps/AccountStep.tsx
