+++ src/views/studio/SignInJoinView/Steps/ExtensionStep.style.ts
