+++ src/views/studio/SignInJoinView/Steps/ExtensionStep.tsx
