+++ src/views/studio/SignInJoinView/Steps/Steps.style.ts
