+++ src/views/studio/SignInJoinView/Steps/TermsStep.style.ts
