+++ src/views/studio/SignInJoinView/Steps/TermsStep.tsx
