+++ src/views/studio/SignInJoinView/Steps/index.ts
