+++ src/views/studio/SignInJoinView/index.ts
@@ -1,0 +1,3 @@
+import SignInJoinView from './SignInJoinView'
+
+export default SignInJoinView
