+++ src/views/studio/index.ts
@@ -2,5 +2,16 @@ import StudioLayout from './StudioLayout'
 import CreateEditChannelView from './CreateEditChannelView'
 import MyVideosView from './MyVideosView'
 import MyUploadsView from './MyUploadsView'
+import SignInView from './SignInView'
+import SignInJoinView from './SignInJoinView'
+import CreateMemberView from './CreateMemberView'
 
-export { StudioLayout, CreateEditChannelView, MyVideosView, MyUploadsView }
+export {
+  StudioLayout,
+  CreateEditChannelView,
+  MyVideosView,
+  MyUploadsView,
+  SignInView,
+  SignInJoinView,
+  CreateMemberView,
+}
